+++ PowerSync/PowerSync.Common/DB/Schema/Attributes/IndexAttribute.cs
@@ -1,0 +1,15 @@
+namespace PowerSync.Common.DB.Schema.Attributes;
+
+[AttributeUsage(AttributeTargets.Class, AllowMultiple = true)]
+public class IndexAttribute : Attribute
+{
+    public string Name { get; }
+    public string[] Columns { get; }
+
+    public IndexAttribute(string name, string[] columns)
+    {
+        Name = name;
+        Columns = columns;
+    }
+}
+